Spawner, Enemigos casi acabados y torretas acbaadas a falta de metodos de recarga y cura

+++ GameJam-2020/Assets/Code/Classes/IDamage.cs
@@ -1,3 +1,3 @@
 public interface IDamage{
-    void ReceiveDamage(float damageAmount);
+    void ReceiveDamage(int damageAmount);
 }
